+++ src/upload_plugg/__init__.py
@@ -1,5 +1,5 @@
 """UPLOAD PLUGG application package."""
 
 APP_NAME = "UPLOAD PLUGG"
-APP_VERSION = "1.0.1"
+APP_VERSION = "1.0.2"
 CREATOR_CREDIT = "Powered by: Dakuza"